+++ src/main/kotlin/cloud/glitchdev/rfu/config/categories/JerryFishing.kt
@@ -11,4 +11,16 @@ object JerryFishing : Category("Jerry Fishing") {
         name = Literal("Reindrake Alert")
         description = Literal("Sends an alert when someone summons a Reindrake!")
     }
+
+    var blizzardTimerDisplay by observable(boolean(true) {
+        name = Literal("Blizzard Timer Display")
+        description = Literal("Display the current blizzard timer on screen")
+    }) { _, _ ->
+        reloadScreen()
+    }
+
+    var blizzardExpiredAlert by boolean(true) {
+        name = Literal("Blizzard Expired Alert")
+        description = Literal("Sends an alert whenever the blizzard expires.")
+    }
 }